Move theme-related files under theme folder
