+++ packages/ra-ui-materialui/src/button/ToggleThemeLegacyButton.tsx
@@ -3,8 +3,8 @@ import { Tooltip, IconButton } from '@mui/material';
 import Brightness4Icon from '@mui/icons-material/Brightness4';
 import Brightness7Icon from '@mui/icons-material/Brightness7';
 import { useTranslate } from 'ra-core';
-import { useTheme } from '../layout';
-import { RaThemeOptions } from '..';
+
+import { useTheme, RaThemeOptions } from '../theme';
 
 /**
  * Button toggling the theme (light or dark).
