--- packages/ra-ui-materialui/src/layout/Theme/index.ts
@@ -1,5 +1,0 @@
-export * from './useTheme';
-export * from './ThemeProvider';
-export * from './ThemesContext';
-export * from './useThemesContext';
-export * from './types';
