+++ packages/ra-ui-materialui/src/layout/index.ts
@@ -21,7 +21,6 @@ export * from './ResourceMenuItem';
 export * from './ResourceMenuItems';
 export * from './Sidebar';
 export * from './SidebarToggleButton';
-export * from './Theme';
 export * from './Title';
 export * from './TitlePortal';
 export * from './TopToolbar';
