+++ packages/ra-ui-materialui/src/theme/index.ts
@@ -1,2 +1,7 @@
+export * from './useTheme';
+export * from './ThemeProvider';
+export * from './ThemesContext';
+export * from './useThemesContext';
+export * from './types';
 export * from './defaultTheme';
 export * from './nanoTheme';
